todomvc/requirements.txt
django
djangorestframework
django-jet
django-reversion
